+++ src/tools/fep.py
@@ -89,149 +89,6 @@ class FEP:
 
         return fep
 
-    @beartype
-    def _ensure_dbs_for_fep(self):
-        """ensure that there are databases for FEP on the data hub
-
-        Ensures the following DBs:
-
-        - ligand (list of ligand files)
-        - protein (list of protein files)
-        - ABFE (ligand, protein, step, job_id, output, results, delta_g)
-        - RBFE (ligand1, ligand2, protein, step, job_id, output, results, delta_g)
-
-        """
-
-        # ligands
-        database = _ensure_database(DB_LIGANDS)
-        required_columns = [
-            dict(name=COL_LIGAND, type="file"),
-        ]
-        database = _ensure_columns(
-            database=database,
-            required_columns=required_columns,
-        )
-        self._ligands_db = database
-
-        # proteins
-        database = _ensure_database(DB_PROTEINS)
-        required_columns = [
-            dict(name=COL_PROTEIN, type="file"),
-        ]
-        database = _ensure_columns(
-            database=database,
-            required_columns=required_columns,
-        )
-        self._proteins_db = database
-
-        # ABFE
-        database = _ensure_database(DB_ABFE)
-        required_columns = [
-            dict(name=COL_PROTEIN, type="text"),
-            dict(name=COL_LIGAND, type="text"),
-            dict(name=COL_STEP, type="text"),
-            dict(name=COL_JOBID, type="text"),
-            dict(name=COL_OUTPUT, type="file"),
-            dict(name=COL_RESULT, type="file"),
-            dict(name=COL_DELTA_G, type="float"),
-        ]
-        database = _ensure_columns(
-            database=database,
-            required_columns=required_columns,
-        )
-        self._abfe_db = database
-
-        # RBFE
-        database = _ensure_database(DB_RBFE)
-        required_columns = [
-            dict(name=COL_PROTEIN, type="text"),
-            dict(name=COL_LIGAND1, type="text"),
-            dict(name=COL_LIGAND2, type="text"),
-            dict(name=COL_STEP, type="text"),
-            dict(name=COL_JOBID, type="text"),
-            dict(name=COL_OUTPUT, type="file"),
-            dict(name=COL_RESULT, type="file"),
-            dict(name=COL_DELTA_DELTA_G, type="float"),
-        ]
-        database = _ensure_columns(
-            database=database,
-            required_columns=required_columns,
-        )
-        self._rbfe_db = database
-
-    def show_ligands(self):
-        """show ligands in the FEP object in a dataframe. This method visualizes the ligands using core-aligned 2D visualizations. To simply obtain a listing of the ligands in this class, use `.ligands`."""
-
-        # convert SMILES to aligned images
-        smiles_list = [ligand.smiles_string for ligand in self.ligands]
-        id_list = [ligand._do_id for ligand in self.ligands]
-        file_list = [os.path.basename(ligand.file) for ligand in self.ligands]
-
-        images = chemistry.smiles_list_to_base64_png_list(smiles_list)
-
-        data = {
-            "ID": id_list,
-            "File": file_list,
-            "Ligand": images,
-        }
-
-        # find all the common properties in all ligands
-        common_keys = set.intersection(
-            *(set(ligand.properties.keys()) for ligand in self.ligands)
-        )
-        for key in common_keys:
-            data[key] = [ligand.properties[key] for ligand in self.ligands]
-
-        df = pd.DataFrame(data)
-
-        # Use escape=False to allow the <img> tags to render as images
-        from IPython.display import HTML, display
-
-        display(HTML(df.to_html(escape=False)))
-
-    def connect(self):
-        """Connects the local instantiation of the simulation to Deep Origin.
-
-        If contained ligand or protein files do not exist on Deep Origin, they will be uploaded. The connect method also connects to existing runs, if any.
-
-        """
-
-        self._ensure_dbs_for_fep()
-
-        # ensure that ligands are uploaded
-        df = api.get_dataframe(DB_LIGANDS)
-
-        for ligand in self.ligands:
-            ligand_file = os.path.basename(ligand.file)
-            matching_indices = df.index[df[COL_LIGAND] == ligand_file].tolist()
-            if len(matching_indices) == 0:
-                print(f"Uploading {ligand.file}...")
-                response = api.upload_file_to_new_database_row(
-                    database_id=DB_LIGANDS,
-                    column_id=COL_LIGAND,
-                    file_path=str(ligand.file),
-                )
-
-                ligand._do_id = response.rows[0].hid
-            else:
-                ligand._do_id = matching_indices[0]
-
-        # ensure that protein is uploaded
-        df = api.get_dataframe(DB_PROTEINS)
-        protein_file = os.path.basename(self.protein.file)
-        matching_indices = df.index[df[COL_PROTEIN] == protein_file].tolist()
-        if len(matching_indices) == 0:
-            print(f"Uploading {self.protein.file}...")
-            response = api.upload_file_to_new_database_row(
-                database_id=DB_PROTEINS,
-                column_id=COL_PROTEIN,
-                file_path=str(self.protein.file),
-            )
-
-            self.protein._do_id = response.rows[0].hid
-        else:
-            self.protein._do_id = matching_indices[0]
-
     def get_abfe_results(self):
         """Get ABFE results.
 
@@ -332,152 +189,6 @@ class FEP:
         from IPython.display import HTML, display
 
         display(HTML(df.to_html(escape=False)))
-
-    def abfe_end_to_end(
-        self,
-        *,
-        ligand_ids: Optional[str] = None,
-    ) -> str:
-        """Method to run an end-to-end ABFE run.
-
-        Args:
-            ligand_ids (Optional[str], optional): List of ligand IDs to run. Defaults to None. When None, all ligands in the object will be run. To view a list of valid ligand IDs, use the `.show_ligands()` method"""
-
-        if ligand_ids is None:
-            ligand_ids = [ligand._do_id for ligand in self.ligands]
-
-        # check that protein ID is valid
-        if self.protein._do_id is None:
-            raise DeepOriginException(
-                "Protein has not been uploaded yet. Use .connect() first."
-            )
-
-        # check that ligand IDs are valid
-        valid_ligand_ids = [ligand._do_id for ligand in self.ligands]
-
-        if None in valid_ligand_ids:
-            raise DeepOriginException(
-                "Some ligands have not been uploaded yet. Use .connect() first."
-            )
-
-        if not set(ligand_ids).issubset(valid_ligand_ids):
-            raise DeepOriginException(
-                f"Some ligand IDs re not valid. Valid ligand IDs are: {valid_ligand_ids}"
-            )
-
-        database_columns = (
-            self._ligands_db.cols + self._proteins_db.cols + self._abfe_db.cols
-        )
-
-        # only run on ligands that have not been run yet
-        # first check that there are no existing runs
-        df = api.get_dataframe(DB_ABFE)
-        df[df["Protein"] == self.protein._do_id]
-
-        df = df[(df[COL_LIGAND].isin(ligand_ids)) & (~pd.isna(df[COL_RESULT]))]
-
-        already_run_ligands = set(df[COL_LIGAND])
-        ligand_ids = set(ligand_ids) - already_run_ligands
-
-        for ligand_id in ligand_ids:
-            _start_abfe_run_and_log(
-                protein_id=self.protein._do_id,
-                ligand_id=ligand_id,
-                database_columns=database_columns,
-                params=self.params_abfe_end_to_end,
-            )
-
-
-@beartype
-def _start_abfe_run_and_log(
-    *,
-    protein_id: str,
-    ligand_id: str,
-    params: dict,
-    database_columns: list,
-):
-    """starts a single run of ABFE end to end and logs it in the ABFE database. Internal function. Do not use.
-
-    Args:
-        protein_id (str): protein ID
-        ligand_id (str): ligand ID
-        params (dict): parameters for the ABFE end-to-end job
-        database_columns (list): list of database columns dicts
-
-    """
-
-    from deeporigin.tools import run
-
-    tool_key = "deeporigin.abfe-end-to-end"
-
-    # make a new row
-    response = api.make_database_rows(DB_ABFE, n_rows=1)
-    row_id = response.rows[0].hid
-
-    # write ligand ID
-    api.set_cell_data(
-        ligand_id,
-        column_id=COL_LIGAND,
-        row_id=row_id,
-        database_id=DB_ABFE,
-    )
-
-    # write protein ID
-    api.set_cell_data(
-        protein_id,
-        column_id=COL_PROTEIN,
-        row_id=row_id,
-        database_id=DB_ABFE,
-    )
-
-    # write step
-    api.set_cell_data(
-        "End-to-end",
-        column_id=COL_STEP,
-        row_id=row_id,
-        database_id=DB_ABFE,
-    )
-
-    # start job
-    params["protein"] = {
-        "columnId": COL_PROTEIN,
-        "rowId": protein_id,
-        "databaseId": DB_PROTEINS,
-    }
-
-    params["ligand"] = {
-        "columnId": COL_LIGAND,
-        "rowId": ligand_id,
-        "databaseId": DB_LIGANDS,
-    }
-
-    outputs = {
-        "output_file": {
-            "columnId": COL_OUTPUT,
-            "rowId": row_id,
-            "databaseId": DB_ABFE,
-        },
-        "abfe_results_summary": {
-            "columnId": COL_RESULT,
-            "rowId": row_id,
-            "databaseId": DB_ABFE,
-        },
-    }
-
-    job_id = run._process_job(
-        inputs=params,
-        outputs=outputs,
-        tool_key=tool_key,
-        cols=database_columns,
-    )
-
-    # write job ID
-    api.set_cell_data(
-        job_id,
-        column_id=COL_JOBID,
-        row_id=row_id,
-        database_id=DB_ABFE,
-    )
 
 
 @beartype
